Got rid of old master gulp file.
