+++ packages/app/main/package.json
@@ -8,14 +8,14 @@
   "homepage": "https://github.com/Microsoft/BotFramework-Emulator",
   "scripts": {
     "build": "run-s lint build:electron",
-    "build:electron": "babel ./src --out-dir app/server --extensions \".ts,.tsx\" --ignore \"**/*.spec.ts\" && gulp copy-extension-stubs",
+    "build:electron": "babel ./src --out-dir app/server --extensions \".ts,.tsx\" --ignore \"**/*.spec.ts\" && gulp --gulpfile gulpfile.common.js copy-extension-stubs",
     "lint": "tslint --project tsconfig.json",
     "start": "concurrently --kill-others --names \"electron,react-app\" --success first \"npm run start:electron:dev\" \"npm run start:react-app\"",
     "start:electron": "./node_modules/.bin/electron --inspect=7777 .",
     "start:electron:dev": "cross-env ELECTRON_TARGET_URL=http://localhost:3000/ npm run start:electron",
     "start:react-app": "cd ../client && npm start",
     "test": "jest",
-    "verify": "gulp verify:copyright"
+    "verify": "gulp --gulpfile gulpfile.common.js verify:copyright"
   },
   "keywords": [
     "microsoft",
